@@ -29,7 +29,7 @@ classifiers = [
 ]
 dynamic = ["version"]
 
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 dependencies = [
     "numpy~=1.20",
     "scipy>=1.5.3, <2",
